+++ R/zzz.R
@@ -2,6 +2,6 @@ utils::globalVariables(".data")
 
 
 .onLoad <- function(libname, pkgname) {
-
+ S7::methods_register()
 }